requirements.txt
Pillow>=10.0.0
pillow-heif>=0.13.0
tqdm>=4.66.0
streamlit>=1.28.0
pandas>=2.0.0

